pyproject.toml
[project]
name = "ai-cookbook"
version = "0.1.0"
requires-python = ">3.7"
readme = "README.md"
authors = [{ name = "Team", email = "team@team.com" }]

dependencies = [
  # Data processing libraries
  "pandas",
  "polars",
  "duckdb==0.9.2",
  # Api server libraries
  "fastapi",
  "typer",
  "uvicorn",
  # Database libraries
  "alembic",
  "pgvector",
  "psycopg[binary]",
  "sqlalchemy",
  # Project libraries
  "arxiv",
  "openai",
  "pypdf",
  "streamlit",
  "matplotlib",
  "plotly",
  "seaborn",
  "scipy",
  "tiktoken",
  "beautifulsoup4",
  "types-beautifulsoup4",
  "pillow",
  "pandas-stubs",
  "types-Pillow",
  "wikipedia",
  # Type checking
  "mypy",
  # Testing
  "pytest",
  # Linting and Formatting
  "ruff",
  # phidata
  "phidata==2.2.0",
]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]

# Change this value to use a different directory for the phidata workspace.
# [tool.phidata]
# workspace = "workspace"

[tool.ruff]
line-length = 110
extend-include = ["*.ipynb"]
exclude = ["aienv*", ".venv*"]
[tool.ruff.per-file-ignores]
# Ignore `F401` (import violations) in all `__init__.py` files
"__init__.py" = ["F401"]
"*.ipynb" = ["E402"]

[tool.mypy]
check_untyped_defs = true
no_implicit_optional = true
warn_unused_configs = true
plugins = ["pydantic.mypy"]
exclude = ["aienv*", ".venv*"]

[[tool.mypy.overrides]]
module = ["pgvector.*", "setuptools.*"]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = "tests"
